Adds new locator for more options menu
