+++ run/test/specs/disappear_after_send_off_1o1.spec.ts
@@ -6,6 +6,7 @@ import {
   DisappearModes,
   USERNAME,
 } from '../../types/testing';
+import { ConversationSettings } from './locators/conversation';
 import {
   DisableDisappearingMessages,
   DisappearingMessagesMenuOption,
@@ -52,7 +53,7 @@ async function disappearAfterSendOff1o1(platform: SupportedPlatformsType) {
   );
 
   // Turn off disappearing messages on device 1
-  await device1.clickOnElementAll({ strategy: 'accessibility id', selector: 'More options' });
+  await device1.clickOnElementAll(new ConversationSettings(device1));
   await device1.clickOnElementAll(new DisappearingMessagesMenuOption(device1));
   await device1.clickOnElementAll(new DisableDisappearingMessages(device1));
   await device1.clickOnElementAll(new SetDisappearMessagesButton(device1));
