+++ run/test/specs/locators/index.ts
@@ -332,23 +332,7 @@ export class InviteContactsMenuItem extends LocatorsInterface {
   }
 }
 
-export class LeaveGroupButton extends LocatorsInterface {
-  public build(): StrategyExtractionObj {
-    switch (this.platform) {
-      case 'android':
-        return {
-          strategy: 'id',
-          selector: `network.loki.messenger:id/title`,
-          text: 'Leave group',
-        };
-      case 'ios':
-        return {
-          strategy: 'accessibility id',
-          selector: 'Leave group',
-        };
-    }
-  }
-}
+
 
 export class DeleteMessageLocally extends LocatorsInterface {
   public build(): StrategyExtractionObj {
